Split all of the Static Audio code into separate classes for Music and Sounds
