+++ src/main/java/org/runejs/client/io/Buffer.java
@@ -18,7 +18,6 @@ public class Buffer extends Node {
     public static int cameraOffsetX = 0;
     public static int lastItemDragTime = 0;
 
-    public static int anInt1982 = 0;
     public static int[] anIntArray1984 = new int[2000];
     public static int anInt1985 = -1;
     public static int anInt1987 = 0;
